requirements.txt
# Data & API
python-binance==1.0.19
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1

# Technical Indicators
ta==0.11.0

# ML Models (Prophet only, no TF)
prophet==1.1.5
# tensorflow removed
# scikit-learn removed (unless needed by prophet dependencies, usually fine)

# API
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Utils
joblib==1.3.2
